GetTickCount: A fix which was recommended by the compiler
git-svn-id: http://svn.freepascal.org/svn/lazarus/trunk@33387 4005530d-fff6-0310-9dd1-cebe43e6787f

+++ components/lazutils/lazutf8sysutils.pas
@@ -102,7 +102,7 @@ var
   tp: TTimeVal;
 begin
   fpgettimeofday(@tp, nil);
-  Result := (tp.tv_sec * 1000) + (tp.tv_usec div 1000);
+  Result := (Int64(tp.tv_sec) * 1000) + (tp.tv_usec div 1000);
 end;
 {$else}
 // Not Windows and not UNIX, so just write the most trivial code until we have something besser:
